+++ src/test/java/io/github/poshjosh/ratelimiter/web/spring/uri/AbstractResourcePathTestBase.java
@@ -11,21 +11,21 @@ abstract class AbstractResourcePathTestBase {
 
     private static final Logger LOG = LoggerFactory.getLogger(AbstractResourcePathTestBase.class);
 
-    abstract ResourcePath<String> givenResourcePath(String... uris);
+    abstract ResourcePath givenResourcePath(String... uris);
 
     @Test
     public void shouldCombine() {
         LOG.debug("shouldCombine()");
-        ResourcePath<String> resourcePath = givenResourcePath("/numbers");
-        ResourcePath<String> result =  resourcePath.combine(givenResourcePath("/1/**", "/2/*"));
-        ResourcePath<String> expected = givenResourcePath("/numbers/1/**", "/numbers/2/*");
+        ResourcePath resourcePath = givenResourcePath("/numbers");
+        ResourcePath result =  resourcePath.combine(givenResourcePath("/1/**", "/2/*"));
+        ResourcePath expected = givenResourcePath("/numbers/1/**", "/numbers/2/*");
         assertThat(result.getPatterns()).isEqualTo(expected.getPatterns());
     }
 
     @Test
     public void shouldMatchDoubleAsterix() {
         LOG.debug("shouldMatchDoubleAsterix()");
-        ResourcePath<String> resourcePath = givenResourcePath("/**");
+        ResourcePath resourcePath = givenResourcePath("/**");
         assertThat( resourcePath.matches("/numbers")).isTrue();
         assertThat( resourcePath.matches("/numbers/1")).isTrue();
     }